Move the 'thread-local stack' implementation up from datastore.
Preparing to tackle #15, which needs similar support in storage.
