+++ gcloud/test__localstack.py
@@ -1,0 +1,43 @@
+# Copyright 2014 Google Inc. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import unittest2
+
+
+class Test__LocalStack(unittest2.TestCase):
+
+    def _getTargetClass(self):
+        from gcloud._localstack import _LocalStack
+
+        return _LocalStack
+
+    def _makeOne(self):
+        return self._getTargetClass()()
+
+    def test_it(self):
+        batch1, batch2 = object(), object()
+        batches = self._makeOne()
+        self.assertEqual(list(batches), [])
+        self.assertTrue(batches.top is None)
+        batches.push(batch1)
+        self.assertTrue(batches.top is batch1)
+        batches.push(batch2)
+        self.assertTrue(batches.top is batch2)
+        popped = batches.pop()
+        self.assertTrue(popped is batch2)
+        self.assertTrue(batches.top is batch1)
+        self.assertEqual(list(batches), [batch1])
+        popped = batches.pop()
+        self.assertTrue(batches.top is None)
+        self.assertEqual(list(batches), [])
